requirements.txt
selenium
pyautogui
pyinstaller
keyboard
psycopg2-binary
python-dotenv
openai
tiktoken
